move persistence action types to actions file
