+++ app/state/actions/persistence.js
@@ -1,10 +1,11 @@
-export const PERSISTENCE_LOAD_STATE_REQUEST = 'PERSISTENCE:LOAD_STATE_REQUEST'
-export const PERSISTENCE_LOAD_STATE_SUCCESS = 'PERSISTENCE:LOAD_STATE_SUCCESS'
-export const PERSISTENCE_LOAD_STATE_FAILURE = 'PERSISTENCE:LOAD_STATE_FAILURE'
-
-export const PERSISTENCE_SAVE_STATE_REQUEST = 'PERSISTENCE:SAVE_STATE_REQUEST'
-export const PERSISTENCE_SAVE_STATE_SUCCESS = 'PERSISTENCE:SAVE_STATE_SUCCESS'
-export const PERSISTENCE_SAVE_STATE_FAILURE = 'PERSISTENCE:SAVE_STATE_FAILURE'
+import {
+  PERSISTENCE_LOAD_STATE_FAILURE,
+  PERSISTENCE_LOAD_STATE_REQUEST,
+  PERSISTENCE_LOAD_STATE_SUCCESS,
+  PERSISTENCE_SAVE_STATE_FAILURE,
+  PERSISTENCE_SAVE_STATE_REQUEST,
+  PERSISTENCE_SAVE_STATE_SUCCESS
+} from '@state/actions'
 
 export const loadStateRequest = () => {
   return {
