+++ app/state/reducers/profile/verification.js
@@ -3,7 +3,7 @@ import {
   VERIFICATION_SUBMIT_CODE_SUCCESS,
   VERIFICATION_SUBMIT_CODE_FAILURE
 } from '@state/actions/profile'
-import { PERSISTENCE_LOAD_STATE_SUCCESS } from '@state/actions/persistence'
+import { PERSISTENCE_LOAD_STATE_SUCCESS } from '@state/actions'
 
 const initialState = {
   isLoading: false,
